マイグレーション&シーダー実行

+++ database/migrations/2022_03_18_090359_create_notices_table.php
@@ -15,7 +15,7 @@ class CreateNoticesTable extends Migration
     {
         Schema::create('notices', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->bigIncrements('team_id')->unsigned();
+            $table->integer('team_id')->unsigned();
         });
     }
