+++ database/migrations/2022_03_18_090432_create_time__lines_table.php
@@ -15,7 +15,7 @@ class CreateTimeLinesTable extends Migration
     {
         Schema::create('time__lines', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->bigIncrements('team_id')->unsigned();
+            $table->integer('team_id')->unsigned();
         });
     }
