+++ database/migrations/2022_03_18_091146_create_posts_table.php
@@ -17,8 +17,8 @@ class CreatePostsTable extends Migration
             $table->bigIncrements('id');
             $table->string('Title', 50);
             $table->string('Body', 2000);
-            $table->bigIncrements('user_id')->unsigned();
-            $table->bigIncrements('file_id')->unsigned();
+            $table->integer('user_id')->unsigned();
+            $table->integer('file_id')->unsigned();
         });
     }
