+++ database/migrations/2022_03_18_091322_create_time__line__posts_table.php
@@ -17,8 +17,8 @@ class CreateTimeLinePostsTable extends Migration
             $table->bigIncrements('id');
             $table->string('Title', 50);
             $table->string('Body', 2000);
-            $table->bigIncrements('user_id')->unsigned();
-            $table->bigIncrements('time_line_id')->unsigned();
+            $table->integer('user_id')->unsigned();
+            $table->integer('time_line_id')->unsigned();
         });
     }
