+++ database/migrations/2022_03_18_091603_create_comments_table.php
@@ -16,8 +16,8 @@ class CreateCommentsTable extends Migration
         Schema::create('comments', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->string('Body', 500);
-            $table->bigIncrements('user_id')->unsigned();
-            $table->bigIncrements('post_id')->unsigned();
+            $table->integer('user_id')->unsigned();
+            $table->integer('post_id')->unsigned();
         });
     }
