+++ database/migrations/2022_03_18_091621_create_goods_table.php
@@ -15,8 +15,8 @@ class CreateGoodsTable extends Migration
     {
         Schema::create('goods', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->bigIncrements('user_id')->unsigned();
-            $table->bigIncrements('post_id')->unsigned();
+            $table->integer('user_id')->unsigned();
+            $table->integer('post_id')->unsigned();
         });
     }
